Allow bilmur to work with WP Rocket active
The WP Rocket feature `Delay JavaScript Execution` prevents the bilmur
beacon with the performance data from being sent.  This looks to see if
the WP Rocket plugin is active and when it is adds the `nowprocket`
attribute to the bilmur script tag.  This tells WP Rocket to skip trying
to delay running this script, which allows use to still collect performance
data.